requirements.txt
asgiref==3.3.4
Django==3.2.4
django-filter==2.4.0
Pillow==8.2.
django-allauth==0.44.0
djangorestframework==3.12.4


